Steps toward funneling unsubbed folk through checkout
