+++ src/components/StripeSessionParamWrapper.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { Navigate, Outlet, useSearchParams } from 'react-router-dom'
 
-const StateParamWrapper = () => {
+const StripeSessionParamWrapper = () => {
   const [search] = useSearchParams()
   const stripeSessionParam = search.get('session_id')
 
   return stripeSessionParam ? <Outlet /> : <Navigate to="/" replace />
 }
 
-export default StateParamWrapper
+export default StripeSessionParamWrapper
